% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwalk.R
\name{compare_pulse}
\alias{compare_pulse}
\title{Title}
\usage{
compare_pulse(dat, fil, vmax, km, pulses, pulse_freq, release, bin_size,
  electrode_distance, dead_space_distance, diffusion_coefficient,
  smoothing_count, convert_current, calibration_current = NULL,
  calibration_concentration = NULL)
}
\arguments{
\item{dat}{}

\item{fil}{}

\item{vmax}{}

\item{km}{}

\item{pulses}{}

\item{pulse_freq}{}

\item{release}{}

\item{bin_size}{}

\item{electrode_distance}{}

\item{dead_space_distance}{}

\item{diffusion_coefficient}{}

\item{smoothing_count}{}

\item{convert_current}{}

\item{calibration_current}{}

\item{calibration_concentration}{}
}
\description{
Title
}
