% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwalk.R
\name{compare_pulse_args_df}
\alias{compare_pulse_args_df}
\title{Compare random walk to experimental data (arguments in data frame)}
\usage{
compare_pulse_args_df(dat, fil, args_df)
}
\arguments{
\item{dat}{Data frame. Experimental data for one stimulus supplied by
read_experiment_csv().}

\item{fil}{Character. String representing the file which supplied the
dat data frame. For annotation of the plot.}

\item{args_df}{One-row data frame. Arguments for calculating a random
walk.}
}
\description{
A wrapper for compare_pulse, supporting an alternate protocol with
the random walk arguments supplied in a one-row data frame.
}
