% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwalk.R
\name{electrode_pos}
\alias{electrode_pos}
\title{Electrode position}
\usage{
electrode_pos(rw, time_column = TRUE)
}
\arguments{
\item{rw.}{A calculated random walk matrix in the form of a data frame.
The electrode bin is the middle column of the data frame.}

\item{time_column.}{Logical. Indicates the presence of a time series column
at the front of the data frame.}
}
\value{
Numeric. The column index indicating the location in the data frame
of the electrode results.
}
\description{
The index of the column in the random walk matrix representing the location
of the electrode. The model locates the electrode in the middle of a
one-dimensional string of release points. However, for performance
optimization, the release bins to the right of the electrode are not
populated. They are not needed for the calculation of the electrode results,
since the bin values to the right and left of the electrode are symmetric.
}
