% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwalk.R
\name{electrode_results}
\alias{electrode_results}
\title{Electrode results}
\usage{
electrode_results(rwalk_df, electrode_pos, smoothing_count = 4)
}
\arguments{
\item{rwalk_df.}{A calculated random walk matrix in the form of a
data frame.}

\item{electrode_pos.}{The column index within the random walk matrix,
locating the electrode results.}

\item{smoothing_count.}{Integer. The electrode results can be smoothed
using a moving average applied across the indicated number of values.}
}
\description{
Returns the time series and electrode results of a given random walk
matrix. The rest of the matrix serves only to compute the electrode
results and are unnecessary for subsequent plotting.
}
