% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwalk.R
\name{get_best_args}
\alias{get_best_args}
\title{Get arguments for best fit}
\usage{
get_best_args(arg_df)
}
\arguments{
\item{arg_df}{Data frame. Calculated by calc_fit_multi(). One row contains
the arguments for building a random walk model. Column arg_df$r2 is the
r-squared value correlating the model built by those arguments and the experimental
data.}
}
\value{
One-row data frame.
}
\description{
Given a data frame of arguments for building a random walk model, return
the ones that fit best to the experimental data determined by their r-squared
value.
}
