% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwalk.R
\name{iteration_duration}
\alias{iteration_duration}
\title{Iteration duration}
\usage{
iteration_duration(diffusion_coefficient, bin_size)
}
\arguments{
\item{diffusion_coefficient}{Square centimeters per second.}

\item{bin_size}{Distance between release bins in microns.}
}
\value{
numeric
}
\description{
Return the length of time between iterations of a random walk matrix.
Given by: t = x^2 / 2D
}
