% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwalk.R
\name{plot_rwalk_compare}
\alias{plot_rwalk_compare}
\title{Title}
\usage{
plot_rwalk_compare(dat_w_src, fil, release, vmax, km, r2,
  calibration_current = NULL, calibration_concentration = NULL)
}
\arguments{
\item{dat_w_src}{}

\item{fil}{}

\item{release}{}

\item{vmax}{}

\item{km}{}

\item{r2}{}

\item{calibration_current}{}

\item{calibration_concentration}{}
}
\description{
Title
}
