% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwalk.R
\name{rwalk_cv_pulse}
\alias{rwalk_cv_pulse}
\title{Build a random walk matrix for cyclic voltammetry}
\usage{
rwalk_cv_pulse(vmax, km, release, pulses, pulse_freq, bin_size,
  electrode_distance, dead_space_distance, diffusion_coefficient, duration)
}
\arguments{
\item{vmax}{For Michaelis-Menten correction of uptake. Microns per second.}

\item{km}{For Michaelis-Menten. Microns.}

\item{release}{Concentration in micromoles.}

\item{pulses}{Count of pulses in the pulse train.}

\item{pulse_freq}{Pulse frequency in Hz.}

\item{bin_size}{Spacing of the modelled release bins in microns.}

\item{electrode_distance}{Distance from the electrode to the last bin in microns.}

\item{dead_space_distance}{Distance in microns of one side of the dead space measured
from the electrode.}

\item{diffusion_coefficient}{Diffusion constant in square centimeters per second.}

\item{duration}{Span of time for modelling release diffusion measured in seconds.}
}
\value{
rw_df. Data frame of the modelled random walk. The first column is
rw_df$time_sec, the time in seconds of the iteration. The results reside in the
values modelled at the location of the electrode in column rw_df$electrode.
}
\description{
The random walk models concentration at time intervals and bin distances
from an electrode. For performance optimization, only bins to the left of
the electrode are included in the matrix, as the right bin values are
symmetric with respect to the left.
}
\details{
Given a pulse train, the function prorates the given release evenly to the
nearest appropriate time iteration.

Resolution of the matrix can be customized with the parameters bin_size and
electrode_distance.

The function supports an assumption of dead space at the insertion point
of the electrode.
}
