% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwalk.R
\name{micmen}
\alias{micmen}
\title{Title}
\usage{
micmen(x, vmax = 4.57, km = 0.78, duration)
}
\arguments{
\item{x}{numeric Concentration in micro-Molar.}

\item{vmax}{numeric in micro-Molar/s.}

\item{km}{numeric in micro-Molar.}

\item{duration}{numeric in seconds.}
}
\value{
numeric
}
\description{
Title
}
\examples{
micmen(x = 1.375, vmax = 4.57, duration = .007407)
}
